+++ test/_utility/fs.ts
@@ -1,8 +1,10 @@
-import { mkdir, rm, writeFile } from "node:fs/promises";
+import { mkdir, readFile, rm, writeFile } from "node:fs/promises";
 import { join, relative, resolve } from "node:path";
 
 import { ExecutionContext } from "ava";
 
+import { Source } from "../../src/source.js";
+import { sourceIdToFilename } from "../../src/source-id.js";
 import { unindent } from "./unindent.js";
 
 const packageRoot = resolve(__dirname, "../../..");
@@ -38,6 +40,31 @@ export async function createFsLayout(moduleFilename: string, t: ExecutionContext
 		}
 	})(dirname, content);
 	return dirname;
+}
+
+export async function verifyFsLayout(t: ExecutionContext, dirname: string, content: FsLayout): Promise<void> {
+	await (async function verify(filename: string, content: string | FsLayout) {
+		if (typeof content === "string") {
+			t.is(await readFile(filename, "utf-8"), content);
+		} else {
+			for (const name in content) {
+				await verify(join(filename, name), content[name]);
+			}
+		}
+	})(dirname, content);
+}
+
+export async function persistSourceUpdates(dirname: string, modifiedSources: Map<string, Source.PersistUpdateCallback>): Promise<void> {
+	for (const [sourceId, persist] of modifiedSources) {
+		const filename = sourceIdToFilename(dirname, sourceId);
+		await persist(filename);
+	}
+}
+
+export async function verifySourceUpdates(moduleFilename: string, t: ExecutionContext, modifiedSources: Map<string, Source.PersistUpdateCallback>, content: FsLayout): Promise<void> {
+	const cwd = await createTempDir(moduleFilename, t);
+	await persistSourceUpdates(cwd, modifiedSources);
+	await verifyFsLayout(t, cwd, content);
 }
 
 export interface FsLayout {